added C++ Program to divide two numbers
Former-commit-id: 2b6f9a1e6b3a85845bda72fffdc7cede17b96343

+++ programme/divide-two-numbers/divide-two-numbers.cpp
@@ -14,7 +14,7 @@ double Divide(int num1, int num2)
 int main() {
     int num1, num2;
     cin>>num1>>num2;
-    int result = Divide(num1,num2);
+    double result = Divide(num1,num2);
     if(result==-1)
     {
         cout<<"Divide by Zero is not possible";
